Tests: Fix inconsistency in testing and actual database schemas.
This bug is extremely annoying --- rails does not update the tester's
view of the database so after a migration it does not detect a change
in the database schema. This makes migrations extremely painful to
implement and causes the supposed robustness of the database schema to
actually be really fragile.